Move create cursor response fns. to AbstractResponseStore
