+++ pinot-broker/src/main/java/org/apache/pinot/broker/requesthandler/BrokerRequestHandlerDelegate.java
@@ -31,12 +31,10 @@ import org.apache.pinot.common.response.BrokerResponse;
 import org.apache.pinot.common.response.CursorResponse;
 import org.apache.pinot.common.response.PinotBrokerTimeSeriesResponse;
 import org.apache.pinot.common.response.broker.BrokerResponseNative;
-import org.apache.pinot.common.response.broker.CursorResponseNative;
 import org.apache.pinot.common.utils.config.QueryOptionsUtils;
 import org.apache.pinot.common.utils.request.RequestUtils;
 import org.apache.pinot.spi.trace.RequestContext;
 import org.apache.pinot.spi.utils.CommonConstants.Broker.Request;
-import org.apache.pinot.spi.utils.TimeUtils;
 import org.apache.pinot.sql.parsers.SqlNodeAndOptions;
 
 
@@ -50,23 +48,15 @@ public class BrokerRequestHandlerDelegate implements BrokerRequestHandler {
   private final BaseSingleStageBrokerRequestHandler _singleStageBrokerRequestHandler;
   private final MultiStageBrokerRequestHandler _multiStageBrokerRequestHandler;
   private final TimeSeriesRequestHandler _timeSeriesRequestHandler;
-  private final AbstractResponseStore _resultStore;
-  private final String _brokerHost;
-  private final int _brokerPort;
-  private final long _expirationIntervalInMs;
-
+  private final AbstractResponseStore _responseStore;
 
   public BrokerRequestHandlerDelegate(BaseSingleStageBrokerRequestHandler singleStageBrokerRequestHandler,
       @Nullable MultiStageBrokerRequestHandler multiStageBrokerRequestHandler,
-      @Nullable TimeSeriesRequestHandler timeSeriesRequestHandler,
-      String brokerHost, int brokerPort, AbstractResponseStore resultStore, String expirationTime) {
+      @Nullable TimeSeriesRequestHandler timeSeriesRequestHandler, AbstractResponseStore responseStore) {
     _singleStageBrokerRequestHandler = singleStageBrokerRequestHandler;
     _multiStageBrokerRequestHandler = multiStageBrokerRequestHandler;
     _timeSeriesRequestHandler = timeSeriesRequestHandler;
-    _brokerHost = brokerHost;
-    _brokerPort = brokerPort;
-    _resultStore = resultStore;
-    _expirationIntervalInMs = TimeUtils.convertPeriodToMillis(expirationTime);
+    _responseStore = responseStore;
   }
 
   @Override
@@ -160,65 +150,11 @@ public class BrokerRequestHandlerDelegate implements BrokerRequestHandler {
 
   private CursorResponse getCursorResponse(int numRows, BrokerResponse response)
       throws Exception {
-    CursorResponse cursorResponse = createCursorResponse(response);
-
-    long submissionTimeMs = System.currentTimeMillis();
-    // Initialize all CursorResponse specific metadata
-    cursorResponse.setBrokerHost(_brokerHost);
-    cursorResponse.setBrokerPort(_brokerPort);
-    cursorResponse.setSubmissionTimeMs(submissionTimeMs);
-    cursorResponse.setExpirationTimeMs(submissionTimeMs + _expirationIntervalInMs);
-    cursorResponse.setOffset(0);
-    cursorResponse.setNumRows(response.getNumRowsResultSet());
-
     long cursorStoreStartTimeMs = System.currentTimeMillis();
-    _resultStore.storeResponse(cursorResponse);
+    _responseStore.storeResponse(response);
     long cursorStoreTimeMs = System.currentTimeMillis() - cursorStoreStartTimeMs;
-    CursorResponse cursorResponse1 = _resultStore.handleCursorRequest(response.getRequestId(), 0, numRows);
-    cursorResponse1.setCursorResultWriteTimeMs(cursorStoreTimeMs);
-    return cursorResponse1;
-  }
-
-  public static CursorResponseNative createCursorResponse(BrokerResponse response) {
-    CursorResponseNative responseNative = new CursorResponseNative();
-
-    // Copy all the member variables of BrokerResponse to CursorResponse.
-    responseNative.setResultTable(response.getResultTable());
-    responseNative.setNumRowsResultSet(response.getNumRowsResultSet());
-    responseNative.setExceptions(response.getExceptions());
-    responseNative.setNumGroupsLimitReached(response.isNumGroupsLimitReached());
-    responseNative.setTimeUsedMs(response.getTimeUsedMs());
-    responseNative.setRequestId(response.getRequestId());
-    responseNative.setBrokerId(response.getBrokerId());
-    responseNative.setNumDocsScanned(response.getNumDocsScanned());
-    responseNative.setTotalDocs(response.getTotalDocs());
-    responseNative.setNumEntriesScannedInFilter(response.getNumEntriesScannedInFilter());
-    responseNative.setNumEntriesScannedPostFilter(response.getNumEntriesScannedPostFilter());
-    responseNative.setNumServersQueried(response.getNumServersQueried());
-    responseNative.setNumServersResponded(response.getNumServersResponded());
-    responseNative.setNumSegmentsQueried(response.getNumSegmentsQueried());
-    responseNative.setNumSegmentsProcessed(response.getNumSegmentsProcessed());
-    responseNative.setNumSegmentsMatched(response.getNumSegmentsMatched());
-    responseNative.setNumConsumingSegmentsQueried(response.getNumConsumingSegmentsQueried());
-    responseNative.setNumConsumingSegmentsProcessed(response.getNumConsumingSegmentsProcessed());
-    responseNative.setNumConsumingSegmentsMatched(response.getNumConsumingSegmentsMatched());
-    responseNative.setMinConsumingFreshnessTimeMs(response.getMinConsumingFreshnessTimeMs());
-    responseNative.setNumSegmentsPrunedByBroker(response.getNumSegmentsPrunedByBroker());
-    responseNative.setNumSegmentsPrunedByServer(response.getNumSegmentsPrunedByServer());
-    responseNative.setNumSegmentsPrunedInvalid(response.getNumSegmentsPrunedInvalid());
-    responseNative.setNumSegmentsPrunedByLimit(response.getNumSegmentsPrunedByLimit());
-    responseNative.setNumSegmentsPrunedByValue(response.getNumSegmentsPrunedByValue());
-    responseNative.setBrokerReduceTimeMs(response.getBrokerReduceTimeMs());
-    responseNative.setOfflineThreadCpuTimeNs(response.getOfflineThreadCpuTimeNs());
-    responseNative.setRealtimeThreadCpuTimeNs(response.getRealtimeThreadCpuTimeNs());
-    responseNative.setOfflineSystemActivitiesCpuTimeNs(response.getOfflineSystemActivitiesCpuTimeNs());
-    responseNative.setRealtimeSystemActivitiesCpuTimeNs(response.getRealtimeSystemActivitiesCpuTimeNs());
-    responseNative.setOfflineResponseSerializationCpuTimeNs(response.getOfflineResponseSerializationCpuTimeNs());
-    responseNative.setRealtimeResponseSerializationCpuTimeNs(response.getRealtimeResponseSerializationCpuTimeNs());
-    responseNative.setExplainPlanNumEmptyFilterSegments(response.getExplainPlanNumEmptyFilterSegments());
-    responseNative.setExplainPlanNumMatchAllFilterSegments(response.getExplainPlanNumMatchAllFilterSegments());
-    responseNative.setTraceInfo(response.getTraceInfo());
-
-    return responseNative;
+    CursorResponse cursorResponse = _responseStore.handleCursorRequest(response.getRequestId(), 0, numRows);
+    cursorResponse.setCursorResultWriteTimeMs(cursorStoreTimeMs);
+    return cursorResponse;
   }
 }
